Fixed MaskDataDword() and AndDataDword()
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler@2173 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ DScaler/PCICard.cpp
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// $Id: PCICard.cpp,v 1.7 2002-06-16 18:53:36 robmuller Exp $
+// $Id: PCICard.cpp,v 1.8 2002-09-10 12:13:37 atnak Exp $
 /////////////////////////////////////////////////////////////////////////////
 // Copyright (c) 2001 John Adcock.  All rights reserved.
 /////////////////////////////////////////////////////////////////////////////
@@ -18,6 +18,10 @@
 // CVS Log
 //
 // $Log: not supported by cvs2svn $
+// Revision 1.7  2002/06/16 18:53:36  robmuller
+// Renamed pciGetDeviceConfig() to pciGetDeviceInfo().
+// Implemented pciGetDeviceConfig() and pciSetDeviceConfig().
+//
 // Revision 1.6  2002/02/12 02:29:40  ittarnavsky
 // fixed the hardware info dialog
 //
@@ -301,7 +305,7 @@ void CPCICard::MaskDataWord(DWORD Offset, WORD Data, WORD Mask)
     WriteWord(Offset, Result);
 }
 
-void CPCICard::MaskDataDword(DWORD Offset, WORD Data, WORD Mask)
+void CPCICard::MaskDataDword(DWORD Offset, DWORD Data, DWORD Mask)
 {
     DWORD Result(ReadDword(Offset));
     Result = (Result & ~Mask) | (Data & Mask);
@@ -343,7 +347,7 @@ void CPCICard::AndDataWord(DWORD Offset, WORD Data)
     WriteWord(Offset, Result);
 }
 
-void CPCICard::AndDataDword(DWORD Offset, WORD Data)
+void CPCICard::AndDataDword(DWORD Offset, DWORD Data)
 {
     DWORD Result(ReadDword(Offset));
     Result &= Data;
